fix: enabling notification pop up for the task, comments and projects

+++ app/Livewire/Task/DeleteTask.php
@@ -4,6 +4,9 @@ namespace App\Livewire\Task;
 
 use App\Models\Task;
 use LivewireUI\Modal\ModalComponent;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Notification;
+use App\Notifications\TaskDeletedNotification;
 
 class DeleteTask extends ModalComponent
 {
@@ -16,8 +19,22 @@ class DeleteTask extends ModalComponent
 
     public function delete()
     {
-        $this->task->delete();
-        $this->dispatch('task-deleted');
+        if ($this->task) {
+            $this->task->delete();
+
+            $deleter = Auth::user();
+            $team = $deleter->currentTeam;
+
+            if ($team) {
+                $usersToNotify = $team->allUsers()->where('id', '!=', $deleter->id);
+
+                Notification::send($usersToNotify, new TaskDeletedNotification($this->task, $deleter->name));
+
+                Notification::send($deleter, new TaskDeletedNotification($this->task, $deleter->name));
+            }
+        }
+
+        $this->dispatch('task-deleted')->to('task.task-kanban-board');
         $this->closeModal();
     }
